Update données
Changement des dmg/speed bullet same pour les zombies, j'ai méthodifié le déplacement,

Il reste -> changer le spawn des zombies, métodifier le conflict et phase 1 over

+++ src/models/IPlant.java
@@ -14,9 +14,9 @@ import zombies.Zombie;
 
 	boolean isDead();
 
-	int getY();
+	float getY();
 
-	int getX();
+	float getX();
 	
 	void action(ArrayList<Projectile> myBullet, BordView view, ArrayList<Zombie> myZombies);
